+++ scitos2_mira/src/main.cpp
@@ -1,26 +1,27 @@
-/*
- * SCITOS NODE
- *
- * Copyright (c) 2022-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
- * 
- * This file is part of scitos_mira project.
- * 
- * All rights reserved.
- *
- */
+// Copyright (c) 2024 Alberto J. Tudela Roldán
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-/// ROS
+#include <memory>
+
 #include "rclcpp/rclcpp.hpp"
+#include "scitos2_mira/mira_framework.hpp"
 
-// Scitos Mira
-#include "scitos2_mira/scitos_mira.hpp"
-
-int main(int argc, char** argv){
-	rclcpp::init(argc, argv);
-	rclcpp::executors::SingleThreadedExecutor exe;
-	auto node = std::make_shared<ScitosMira>("scitos_mira");
-	exe.add_node(node->get_node_base_interface());
-	exe.spin();
-	rclcpp::shutdown();
-	return 0;
+int main(int argc, char ** argv)
+{
+  rclcpp::init(argc, argv);
+  auto node = std::make_shared<scitos2_mira::MiraFramework>();
+  rclcpp::spin(node->get_node_base_interface());
+  rclcpp::shutdown();
+  return 0;
 }